[CCUBE-2182][GZ] add jsdocs to schedule
